+++ include/gtirb/IR.hpp
@@ -777,36 +777,30 @@ public:
 
   /// \brief Iterator over blocks.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using block_iterator =
       MergeSortedIterator<Module::block_iterator, BlockAddressLess>;
   /// \brief Range of blocks.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using block_range = boost::iterator_range<block_iterator>;
   /// \brief Sub-range of blocks overlapping an address or range of addreses.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using block_subrange = boost::iterator_range<
       MergeSortedIterator<Module::block_subrange::iterator, BlockAddressLess>>;
   /// \brief Iterator over blocks.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using const_block_iterator =
       MergeSortedIterator<Module::const_block_iterator, BlockAddressLess>;
   /// \brief Range of blocks.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using const_block_range = boost::iterator_range<const_block_iterator>;
   /// \brief Sub-range of blocks overlapping an address or range of addreses.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using const_block_subrange = boost::iterator_range<MergeSortedIterator<
       Module::const_block_subrange::iterator, BlockAddressLess>>;
 
@@ -952,39 +946,33 @@ public:
 
   /// \brief Iterator over \ref CodeBlock objects.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using code_block_iterator =
       MergeSortedIterator<Module::code_block_iterator, AddressLess>;
   /// \brief Range of \ref CodeBlock objects.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using code_block_range = boost::iterator_range<code_block_iterator>;
   /// \brief Sub-range of \ref CodeBlock objects overlapping an address or range
   /// of addreses.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using code_block_subrange = boost::iterator_range<
       MergeSortedIterator<Module::code_block_subrange::iterator, AddressLess>>;
   /// \brief Iterator over \ref CodeBlock objects.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using const_code_block_iterator =
       MergeSortedIterator<Module::const_code_block_iterator, AddressLess>;
   /// \brief Range of \ref CodeBlock objects.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using const_code_block_range =
       boost::iterator_range<const_code_block_iterator>;
   /// \brief Sub-range of \ref CodeBlock objects overlapping an address or range
   /// of addreses.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using const_code_block_subrange = boost::iterator_range<MergeSortedIterator<
       Module::const_code_block_subrange::iterator, AddressLess>>;
 
@@ -1130,39 +1118,33 @@ public:
 
   /// \brief Iterator over \ref DataBlock objects.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using data_block_iterator =
       MergeSortedIterator<Module::data_block_iterator, AddressLess>;
   /// \brief Range of \ref DataBlock objects.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using data_block_range = boost::iterator_range<data_block_iterator>;
   /// \brief Sub-range of \ref DataBlock objects overlapping an address or range
   /// of addreses.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using data_block_subrange = boost::iterator_range<
       MergeSortedIterator<Module::data_block_subrange::iterator, AddressLess>>;
   /// \brief Iterator over \ref DataBlock objects.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using const_data_block_iterator =
       MergeSortedIterator<Module::const_data_block_iterator, AddressLess>;
   /// \brief Range of \ref DataBlock objects.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using const_data_block_range =
       boost::iterator_range<const_data_block_iterator>;
   /// \brief Sub-range of \ref DataBlock objects overlapping an address or range
   /// of addreses.
   ///
-  /// Blocks are yielded in address order, ascending. If two blocks have the
-  /// same address, thier order is not specified.
+  /// Blocks are yielded in address order, ascending.
   using const_data_block_subrange = boost::iterator_range<MergeSortedIterator<
       Module::const_data_block_subrange::iterator, AddressLess>>;